Add files via upload
This is an automation that allows enterprise restaurants to make updates to their Company Order Guides in bulk.  Updating a single Order Guide manually for as few as a single menu item takes precious minutes away from your work day, possibly even hours or days if you are making an overhaul of changes. 

Order Guides for restaurants are used to determine menu item prices for specific locations.  Our larger clients often times have dozens of order guides consisting of hundreds of items. 

This automation allows for easy price changes to be made on a spreadsheet, which is then processed over the script.